# ^(::Float, ::Integer)