--- a/base/math.jl
+++ b/base/math.jl
@@ -911,8 +911,10 @@ function modf(x::Float64)
 end
 
 @inline function ^(x::Float64, y::Float64)
-    yint = round(Int, y)
-    y == yint && return x^yint
+    if isfinite(x)
+        yint = round(Int, y)
+        y == yint && return x^yint
+    end
     z = ccall("llvm.pow.f64", llvmcall, Float64, (Float64, Float64), x, y)
     if isnan(z) & !isnan(x+y)
         throw_exp_domainerror(x)
@@ -920,8 +922,10 @@ end
     z
 end
 @inline function ^(x::Float32, y::Float32)
-    yint = round(Int, y)
-    y == yint && return x^yint
+    if isfinite(x)
+        yint = round(Int, y)
+        y == yint && return x^yint
+    end
     z = ccall("llvm.pow.f32", llvmcall, Float32, (Float32, Float32), x, y)
     if isnan(z) & !isnan(x+y)
         throw_exp_domainerror(x)