uses compensated power by squaring for Float64. For Float32, this just using power by squaring with Float64. .5 ULP accuracy for both types.
```
x32 = randn(Float32, 1024); y32 = similar(x32);
x64 = 10 .* randn(Float64, 1024); y64 = similar(x64);
```
Old timings:
```
@btime @. $y32 = $x32^101;
  52.840 μs (0 allocations: 0 bytes)
@btime @. $y64 = $x64^101);
  66.471 μs (0 allocations: 0 bytes)
```
New timings:
```
@btime @. $y32 = $x32^101;
4.619 μs (0 allocations: 0 bytes)

@btime @. $y64 = $x64^101;
  8.520 μs (0 allocations: 0 bytes)
```
Furthermore, this gets .5 ULP accuracy on all inputs I've tested so far, and is even faster than what I've shown for smaller powers.