@oscardssmith 2021-08-28
@palday I don't have a final answer for error analysis yet, but here's what I have so far.
The worst positive exponent is `2^62-1` (because for all bigger exponents, the answer is `Inf` or `0`).

In this case, the algorithm simplifies to.
```
function mypow2p62m1(x)
    y = 1.0
    xnlo = ynlo = 0.0
    for _ in 1:63
        yn = x*y
        ynlo = fma(x, y , -yn) + muladd(y, xnlo, x*ynlo) # adds error of xnlo*ynlo + eps(x*ynlo) + eps(y*xnlo)
        y = yn
        xn = x * x
        xnlo = muladd(x, 2*xnlo, fma(x, x, -xn))  # adds error of xnlo^2 + eps(x*xnlo)
        x = xn
    end
    return muladd(x, y, muladd(y, xnlo, x*ynlo)) # adds error of xnlo*ynlo + eps(x*ynlo) + eps(y*xnlo)
end
```
Theoretically, we could get an upper bound on this error, by adding up all the terms (at their maximum), but the intuitive answer is that since every time, the loop introduce very roughly `5eps(eps(xn))` of error, I think the maximal relative error (before the final operation when we return 1 result which has an unavoidable .5 ULP) of roughly `c*eps(eps(x))` where c is a constant that should be <1000.