@KristofferC 2021-08-28
Just some references to some previous accuracy discussions w.r.t this function:

https://github.com/JuliaLang/julia/issues/19872, https://github.com/JuliaLang/julia/pull/19890, https://github.com/JuliaLang/julia/issues/2741, https://github.com/JuliaLang/julia/issues/8939