@PallHaraldsson 2021-08-30
FMA is in basically all platforms, CPUs and Nvidia GPUs, since at least 2011 (and in IEEE 754-2008 revision). Can't we just swallow the slowdown in older CPUs, make a policy change to use FMA (not just for this) given the huge upside? Does LLVM generate the slower code, instead of the instruction, with same accuracy for older platforms?

Could we make `@fast_for_old_processors` macro to not perform "poorly" on old platforms, if people are worried (I'm not)?