@palday 2021-08-30
@PallHaraldsson It's not uncommon for big academic compute infrastructure to have processors that are 10 years out of date, but to have a lot of them in a high memory set up. That way you can benefit from massive parallelism and they're not spending millions every couple of years to upgrade.