@oscardssmith 2021-09-10
Adding a triage label to decide whether it's acceptable for to have `Base` that require `fma` hardware for good performance.