@Samuel3008 2021-09-13
The new implementation needs `n == 0 && return one(x)`.