@oscardssmith 2021-09-13
This actually already returns `1.0` for `n==0`. That said, from reading the code, I'm having trouble figuring out why it works.