@PallHaraldsson 2021-09-13
```
returns ` 1.0` 
```

as it should because of `one(x)`, did you read it as `one(n)`? That's not the common case, so I tested and seems faster this way: