@Samuel3008 2021-09-13
> This actually already returns `1.0` for `n==0`. That said, from reading the code, I'm having trouble figuring out why it works.

Are you sure? I'm not at a computer right now so can't check but I'm 99% confident `x^0` gave `x` when I tried it before. (Which aligns with what I expect reading the code)