@JeffBezanson 2021-09-16
From triage:
- Option 1: Get multiversioning for `if have_fma` working very soon (multiversioning pass replaces an intrinsic with a constant based on target).
- Option 2: Only support CPUs with fma
- Option 3: Don't care about performance on non-fma CPUs.

Some of us think a good ordering is to do option 3 first, and whenever 1 is implemented those systems will just get faster.