@oscardssmith 2021-09-17
@Samuel3008 I figured out what was happening. I was hitting literal_pow which was giving different results.