@chriselrod 2021-09-17
Why not compare `y == round(y)`?