@oscardssmith 2021-09-17
I was trying to do this as efficiently as possible. using `y==round(y)` would require 1 round, 1 check on value (abs(y)<2^64), and then a `round(Int,y)`. This way, if it works only needs 1 rounding.