@PallHaraldsson 2021-10-04
I did extensive benchmarking of your version vs mine, and I didn't see a benefit to your extra code for e.g. power of three:

```
@inline function ^(x::Float64, n::Integer)
           # Doesn't seem to have any effect in the original: n == 0 && return one(x)
           x_original_squared = x*x; y = 1.0
           xnlo = ynlo = 0.0
           if n < 0
               rx = inv(x)
               isfinite(x) && (xnlo = fma(x, rx, -1.) * rx)
               x = rx
               n = -n
           end
           n -= 2; # y = 1.0; # Not needed: n==3 && return x*x*x #keep compatability with literal_pow
           while n > 1
               if n&1 > 0
                   yn = x*y
                   ynlo = fma(x, y, -yn) + muladd(y, xnlo, x*ynlo)
                   y = yn
               end
               xn = x * x
               xnlo = muladd(x, 2*xnlo, fma(x, x, -xn))
               x = xn
               n >>= 1
           end
           # I was thinking of moving here: n == 0 && return one(x)
           !isfinite(x) && return x_original_squared*x*y
           return muladd(x_original_squared*x, y, muladd(y, xnlo, x*ynlo))
       end
```