@oscardssmith 2021-10-04
The reason to special case p=3 isn't for performance. It's so that `x^3` returns the same result as `p=3; x^p`.