@oscardssmith 2021-10-04
The reason I added the code here rather than remove the special case for literal_pow is that I there are a lot of people who rely on the fact that `x^3` is really fast, and making the literal version match the results of this would make it noticeably slower. (specifically, it would expand to
```
x2=x*x
x2lo=fma(x, x, -x2)
return fma(x2, x, x*x2lo)
```
which is a 40% increase in instructions.