@KristofferC 2021-10-06
FWIW, it is probably a good idea to run PkgEval on PRs like this that changes very low level functions.